Add fused-multiply-and-add (and sub) support, see long commit message for details
Added new opcodes:
	cFma = x*y + a
	cFms = x*y - a
	cFmma = x*y + a*b
	cFmms = x*y - a*b

And associated optimization rules.

Using this test expression:
	(z*(x*y+z*2)+x*(y+3))*x+(x*y+z)*4

Timings:

       ,-----------------------------------------------------------------------------,
       |                  Parse |  Eval | Eval (O) | Eval (O2) | Optimize | Repeat O.|
,------+------------------------+-------+----------+-----------+----------+----------+
|         float, before   0.871 | 0.061 |    0.062 |     0.060 |    127.0 |    123.9 |
|         float, after    0.881 | 0.054 |    0.046 |     0.047 |    134.7 |    134.5 |
|-------------------------------+-------+----------+-----------+----------+----------|
|        double, before   0.841 | 0.056 |    0.053 |     0.056 |    128.5 |    125.3 |
|        double, after    0.877 | 0.055 |    0.043 |     0.043 |    134.3 |    134.3 |
|-------------------------------+-------+----------+-----------+----------+----------|
|   long double, before   0.988 | 0.108 |    0.105 |     0.105 |    140.3 |    137.7 |
|   long double, after    0.887 | 0.094 |    0.079 |     0.079 |    137.5 |    137.5 |
|-------------------------------+-------+----------+-----------+----------+----------|
|complex double, before   0.861 | 0.100 |    0.092 |     0.092 |    111.7 |    108.5 |
|complex double, after    0.927 | 0.083 |    0.058 |     0.059 |    126.8 |    125.8 |
|-------------------------------+-------+----------+-----------+----------+----------|
|     MpfrFloat, before   5.361 | 0.963 |    1.030 |     1.023 |    170.7 |    173.1 |
|     MpfrFloat, after    5.627 | 0.967 |    0.977 |     0.973 |    194.5 |    196.0 |
'------------------------------------------------------------------------------------'

Conclusion: Evaluation performance improved on all datatypes, but especially on float, double and their complex equivalents, which greatly benefit from the opportunity of hardware fma instructions.

With MpfrFloat, fma/fms/fmma/fmms are done directly using mpfr functions (instead of individual mul/add operations).
However, improvement with Mpfrfloat was smaller, because the mpfr library actually performs these operations individually, which does not give much performance improvement.

Bytecode:

Original, before                                  Original, after
-------------------                               -------------------
Size of stack: 4                                  Size of stack: 4
0000: push Var2  = [0]z                           push Var2 = [0]z
0001: push Var0  = [1]x                           push Var0 = [1]x
0002: push Var1  = [2]y                           push Var1 = [2]y
0003: mul        = [1]x*y                         mul       = [1]x*y
0004: push Var2  = [2]z                           push Var2 = [2]z
0005: dup        \ ...z*1                         dup       \ ...z*1
0006: add        /_[2]z*2                         add       /_[2]z*2
0007: add        = [1]x*y+z*2                     add       = [1]x*y+z*2
0008: mul        = [0]z*(x*y+z*2)                 mul       = [0]z*(x*y+z*2)
0009: push Var0  = [1]x                           push Var0 = [1]x
000a: push Var1  = [2]y                           push Var1 = [2]y
000b: push 3     = [3]3                           push 3    = [3]3
000c: add        = [2]y+3                         add       = [2]y+3
000d: mul        = [1]x*(y+3)                     mul       = [1]x*(y+3)
000e: add        = [0]z*(x*y+z*2)+x*(y+3)         add       = [0]z*(x*y+z*2)+x*(y+3)
000f: push Var0  = [1]x                           push Var0 = [1]x
0010: mul        = [0](z*(x*y+z*2)+x*(y+3))*x     mul       = [0](z*(x*y+z*2)+x*(y+3))*x
0011: push Var0  = [1]x                           push Var0 = [1]x
0012: push Var1  = [2]y                           push Var1 = [2]y
0013: mul        = [1]x*y                         push Var2 = [3]z
0014: push Var2  = [2]z                           fma       = [1](x*y+z)
0015: add        = [1]x*y+z                       push 4    = [2]4
0016: push 4     = [2]4                           mul       = [1](x*y+z)*4
0017: mul        = [1](x*y+z)*4                   add       = [0](z*(x*y+z*2)+x*(y+3))*x+(x*y+z)*4
0018: add        = [0](z*(x*y+z*2)+x*(y+3))*x+(x*y+z)*4

Optimized, before                   | Optimized, after
---------                           | ----------------
Size of stack: 7                    | Size of stack: 10
0000: push Var0  = [0]x             | 0000: push 4    = [0]4
0001: push Var1  = [1]y             | 0001: push Var0 = [1]x
0002: mul        = [0]x*y           | 0002: push Var1 = [2]y
0003: dup        = [1]x*y           | 0003: push Var2 = [3]z
0004: push Var2  = [2]z             | 0004: fma       = [1](x*y+z)
0005: add        = [1]x*y+z         | 0005: push Var0 = [2]x
0006: push 4     = [2]4             | 0006: dup       = [3]x
0007: mul        = [1](x*y+z)*4     | 0007: push Var1 = [4]y
0008: push Var0  = [2]x             | 0008: push 3    = [5]3
0009: dup        = [3]x             | 0009: add       = [4]y+3
000a: push Var1  = [4]y             | 000a: push Var2 = [5]z
000b: push 3     = [5]3             | 000b: push 2    = [6]2
000c: add        = [4]y+3           | 000c: cFetch(5) = [7]z
000d: mul        = [3]x*(y+3)       | 000e: cFetch(3) = [8]x
000e: push Var2  = [4]z             | 0010: push Var1 = [9]y
000f: dup        = [5]z             | 0011: fmma      = [6](2*z+x*y)
0010: dup        \ ...z*1           | 0012: fmma      = [3](x*(y+3)+z*((2*z+x*y)))
0011: add        /_[5]z*2           | 0013: fmma      = [0](4*((x*y+z))+x*((x*(y+3)+z*((2*z+x*y)))))
0012: cFetch(0)  = [6]x*y
0014: add        = [5]z*2+x*y
0015: mul        = [4]z*(z*2+x*y)
0016: add        = [3]x*(y+3)+z*(z*2+x*y)
0017: mul        = [2]x*(x*(y+3)+z*(z*2+x*y))
0018: add        = [1](x*y+z)*4+x*(x*(y+3)+z*(z*2+x*y))

+++ fpoptimizer/grammar.cc
@@ -138,7 +138,7 @@ namespace FPoptimizer_Grammar
     }
 
     template<typename Value_t>
-    void DumpParams(unsigned paramlist, unsigned count, std::ostream& o)
+    void DumpParams(unsigned long paramlist, unsigned count, std::ostream& o)
     {
         for(unsigned a=0; a<count; ++a)
         {
@@ -158,7 +158,7 @@ namespace FPoptimizer_Grammar
 namespace FPoptimizer_Grammar
 {
 #define FP_INSTANTIATE(type) \
-    template void DumpParams<type>(unsigned, unsigned, std::ostream& ); \
+    template void DumpParams<type>(unsigned long, unsigned, std::ostream& ); \
     template void DumpParam<type>(const ParamSpec&, std::ostream& );
     FPOPTIMIZER_EXPLICITLY_INSTANTIATE(FP_INSTANTIATE)
 #undef FP_INSTANTIATE
